tests/requirements.txt
pytest>=7.0.0
psycopg2-binary>=2.9.0